Improve HTML description in icalendar export
So this sent me down a rabbit whole, I really should have skipped, but
here is the gist.

There is no standard way to include HTML in icalendar and many clients
implement it different.
The standard says that the description field should only contain plain
text.

So macOS calendar doesn't bother to parse HTML neither in the
description nor in any other field.

Microsoft Outlook searches for the X-ALT-DESC field for which the mime
can be set to HTML in which case it will display the HTML inside it.
If that field isn't available it falls back to the default description
field as plain text.

Google Calendar throws a HTML parser against the description field and
if it detects HTML it will parse it as such.

So for maximum support across clients we decided to put plain text into
the description field and support Outlooks X-ALT-DESC.