Improve fault resilance during crawling
The Websites we crawl aren't really made for that so sooner or later
our crawling code will break (because the layout changes or maybe
if the website isn't available all together).

Up until now this would have crashed the whole script. With this
commit we now will continue with all other danceschools and there
is now an error in the json export to tell you what went wrong.

So in the case one crawl function crashes we will still be able to
provide a service, even though some events might be missing.

This can be further improved by better (more meaningfull) error
messages (include the danceschool) and even inside a danceschool we
might want to do something similar if we crawl multiple websites.