Redesign website UI
I wasn't quite happy with the initial design.

The colors were a little too bold and the table wasn't quite
the right way to display the events because on mobile the rows
were just too small, resulting in crammed UI.

Now with cards the UI looks a lot cleaner which allows us to
include more information like the end time.

Also I think this feels more cartoonish, which was the look I wanted
to go for in the beginning but didn't quite pull it off.