Update danceschool schedules and information
Some schools do have a different schedule over the summer, but most
scraping already works with that.

For some schools though it is really difficult so it might be best to
ask them if it would be OK if they could provide a calendar etc.

Anyway, really happy with the state the website is now, I think we
should be good for the summer. ^^